# backend-go/src/docs/swagger.yaml
basePath: /api
definitions:
...
  routes.ErrorResponse:
    properties:
      error:
        description: 错误信息
        type: string
    type: object
...
  routes.GenerateCaptchaResponse:
    properties:
      captcha:
        description: 验证码
        type: integer
      message:
        description: 消息
        type: string
    type: object
...
  routes.GetUserResponse:
    properties:
      message:
        description: 消息
        type: string
      user:
        allOf:
        - $ref: '#/definitions/routes.User'
        description: 用户信息
    type: object
  routes.LoginRequest:
    properties:
      email:
        description: 邮箱地址
        type: string
      password:
        description: 密码
        type: string
    type: object
  routes.LoginResponse:
    properties:
      message:
        description: 消息
        type: string
      token:
        description: JWT token
        type: string
      user:
        allOf:
        - $ref: '#/definitions/routes.User'
        description: 用户信息
    type: object
...
  routes.RegisterRequest:
    properties:
      email:
        description: 邮箱地址
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  routes.RegisterResponse:
    properties:
      message:
        description: 消息
        type: string
      token:
        description: JWT token
        type: string
      user:
        allOf:
        - $ref: '#/definitions/routes.User'
        description: 用户信息
    type: object
...
  routes.SendEmailCodeRequest:
    properties:
      email:
        description: 邮箱地址
        type: string
    type: object
  routes.SendEmailCodeResponse:
    properties:
      id:
        description: ID
        type: string
      message:
        description: 消息
        type: string
    type: object
  routes.SendLoginEmailCodeRequest:
    properties:
      email:
        description: 邮箱地址
        type: string
    type: object
  routes.SendLoginEmailCodeResponse:
    properties:
      id:
        description: ID
        type: string
      message:
        description: 消息
        type: string
    type: object
...
  routes.User:
    properties:
      email:
        description: 邮箱地址
        type: string
      id:
        description: 用户ID
        type: string
      username:
        description: 用户名
        type: string
    type: object
host: localhost:5001
info:
  contact: {}
  description: TodoIng task management system API
  title: TodoIng API
  version: "1.0"
paths:
  /auth/captcha:
    get:
      consumes:
      - application/json
      description: 生成图形验证码用于注册/登录验证
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.GenerateCaptchaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 生成验证码
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取访问令牌
      parameters:
      - description: 用户登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 用户登录
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: 获取当前认证用户的信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.GetUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 获取当前用户信息
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 用户注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 用户注册
      tags:
      - auth
  /auth/send-email-code:
    post:
      consumes:
      - application/json
      description: 发送邮箱验证码用于用户注册验证
      parameters:
      - description: 发送邮箱验证码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.SendEmailCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.SendEmailCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 发送注册邮箱验证码
      tags:
      - auth
  /auth/send-login-email-code:
    post:
      consumes:
      - application/json
      description: 发送邮箱验证码用于用户登录验证
      parameters:
      - description: 发送登录邮箱验证码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.SendLoginEmailCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.SendLoginEmailCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 发送登录邮箱验证码
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"